// frontend/src/api/models/LoginTherapistDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface LoginTherapistDTO
 */
export interface LoginTherapistDTO {
  /**
   *
   * @type {string}
   * @memberof LoginTherapistDTO
   */
  email?: string
  /**
   *
   * @type {string}
   * @memberof LoginTherapistDTO
   */
  password?: string
}

/**
 * Check if a given object implements the LoginTherapistDTO interface.
 */
export function instanceOfLoginTherapistDTO(value: object): value is LoginTherapistDTO {
  return true
}

export function LoginTherapistDTOFromJSON(json: any): LoginTherapistDTO {
  return LoginTherapistDTOFromJSONTyped(json, false)
}

export function LoginTherapistDTOFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): LoginTherapistDTO {
  if (json == null) {
    return json
  }
  return {
    email: json['email'] == null ? undefined : json['email'],
    password: json['password'] == null ? undefined : json['password'],
  }
}

export function LoginTherapistDTOToJSON(json: any): LoginTherapistDTO {
  return LoginTherapistDTOToJSONTyped(json, false)
}

export function LoginTherapistDTOToJSONTyped(
  value?: LoginTherapistDTO | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value
  }

  return {
    email: value['email'],
    password: value['password'],
  }
}